feat(ui): Added more ui elements
Three new functionalities are added to GoogleCloudTTS: DelayElement, TextElement, and SoundElement, in order to enhance user interaction. A VoiceConfig.cs file is created to manage various voice configurations, which is applicable for both Text and Sound Elements. Also made necessary changes in MainWindow.axaml and MainWindowViewModel.cs to accommodate these changes and added corresponding new ViewModels and Views for each of these functionalities.