fix(codebase): Removed old frontend
The frontend version of the GoogleCloudTTS application has been removed, as it's no longer going to be supported. All the files related to the frontend are deleted. Any relevant changes can now be concentrated on the backend and shared projects only. Commit also removes frontend application from the solution file.